chore: rename token model files to kebab-case
